requirements.txt
altgraph==0.17.2
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
idna==3.3
macholib==1.16
paramiko==2.11.0
Pillow==9.2.0
pycparser==2.21
pyinstaller==5.2
pyinstaller-hooks-contrib==2022.8
PyNaCl==1.5.0
requests==2.28.1
six==1.16.0
urllib3==1.26.10
